Add event listener for comment button
